Adicionando elementos à rede
Mais botões inclusos na UI que permitem selecionar, arrastar e posicionar elementos na tela, enquanto estiver editando a rede e não simulando.

+++ Petri_Net/Arch.pde
@@ -3,7 +3,7 @@ class Arch
   PVector start, end;
   int weight;
   
-  Arch(int x0, int y0, int x1, int y1, int w)
+  Arch(float x0, float y0, float x1, float y1, int w)
   {
     this.start = new PVector(x0, y0);
     this.end = new PVector(x1, y1);
